+++ src/tooling/piral-cli/src/common/shell.ts
@@ -1,10 +1,11 @@
+import { Agent } from 'https';
 import { progress } from './log';
 import { packageJson, piletJson } from './constants';
 import { readJson, updateExistingJson, writeJson } from './io';
 import { scaffoldFromEmulatorWebsite } from './website';
 import { combinePackageRef, getPackageName, getPackageVersion } from './npm';
 import { dissectPackageName, installNpmPackage, isLinkedPackage } from './npm';
-import { NpmClientType, PackageType, PiralInstanceDetails } from '../types';
+import { NpmClient, PackageType, PiralInstanceDetails } from '../types';
 
 async function updatePiletJson(target: string, appName: string, appDetails: PiralInstanceDetails) {
   const oldContent = await readJson(target, piletJson);
@@ -29,7 +30,7 @@ async function setupPiralInstance(
   hadVersion: boolean,
   rootDir: string,
   sourceVersion: string,
-  npmClient: NpmClientType,
+  npmClient: NpmClient,
 ) {
   if (!isLinkedPackage(sourceName, type, hadVersion, rootDir)) {
     const packageRef = combinePackageRef(sourceName, sourceVersion, type);
@@ -54,13 +55,14 @@ export async function installPiralInstance(
   usedSource: string,
   baseDir: string,
   rootDir: string,
-  npmClient: NpmClientType,
+  npmClient: NpmClient,
+  agent: Agent,
   selected?: boolean,
 ): Promise<string> {
   const [sourceName, sourceVersion, hadVersion, type] = await dissectPackageName(baseDir, usedSource);
 
   if (type === 'remote') {
-    const emulator = await scaffoldFromEmulatorWebsite(rootDir, sourceName);
+    const emulator = await scaffoldFromEmulatorWebsite(rootDir, sourceName, agent);
     const packageName = emulator.name;
     await updatePiletJson(rootDir, packageName, {
       selected,